# backend/src/main/resources/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "经销商分析助手全域业务 API",
    "version": "4.0.0",
    "description": "提供大模型分析经销商业务的核心API。包含【汇总指标】与【全字段明细】。\n\n**大模型（AI）调度规则（必须遵守）：**\n1. **首选汇总**：遇到分析趋势、达成率、对比、宏观统计时，优先调用 `_metrics` 接口获取已计算好的数据。\n2. **按需抓取明细**：只有用户明确要求查看具体流水记录时，才调用 `_details` 接口。\n3. **分页与归纳法则**：当 `_details` 返回大量数据时，AI必须使用 `page` 和 `pageSize` (默认50) 进行分批抓取。"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "本地开发环境"
    },
    {
      "url": "https://your-api-domain.com",
      "description": "生产环境"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Short-lived opaque access token returned by /api/auth/login or /api/auth/refresh."
      }
    },
    "schemas": {
...
      "ReportDraft": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "reportType": { "type": "string", "enum": ["daily", "weekly", "monthly", "topic"] },
          "title": { "type": "string" },
          "language": { "type": "string", "enum": ["zh", "en"] },
          "markdown": { "type": "string" },
          "generatedAt": { "type": "string", "format": "date-time" },
          "importBatchId": { "type": "string" },
          "scope": {
            "type": "object",
            "properties": {
              "type": { "type": "string", "enum": ["GLOBAL", "ORGANIZATION"] },
              "id": { "type": "string" }
            }
          },
          "model": { "type": "string" },
          "promptVersion": { "type": "string" }
        }
      },
      "ReportDraftRequest": {
        "type": "object",
        "required": ["reportType", "language"],
        "properties": {
          "reportType": { "type": "string", "enum": ["daily", "weekly", "monthly", "topic"] },
          "language": { "type": "string", "enum": ["zh", "en"] },
          "scopeType": { "type": "string", "default": "GLOBAL", "description": "兼容字段；认证请求的实际范围始终由服务端组织 grant 决定。" },
          "scopeId": { "type": "string", "description": "不作为可信授权范围；服务端不会用它扩大当前用户范围。" },
          "topic": { "type": "string", "maxLength": 500 }
        }
      },
      "ReportDraftResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/ReportDraft" }
        }
      },
      "ReportDraftListResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "message": { "type": "string" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReportDraft" }
          }
        }
      },
...
      "OrganizationNodeRequest": {
        "type": "object",
        "required": ["nodeKey", "displayName", "nodeType", "enabled"],
        "properties": {
          "nodeKey": { "type": "string", "pattern": "^[A-Z0-9:_-]{2,128}$" },
          "displayName": { "type": "string", "maxLength": 128 },
          "nodeType": { "type": "string", "enum": ["GROUP", "REGION", "CITY", "DEALER"] },
          "parentId": { "type": "integer", "format": "int64", "nullable": true },
          "enabled": { "type": "boolean" }
        }
      },
      "OrganizationNodeUpdateRequest": {
        "type": "object",
        "required": ["displayName", "nodeType", "enabled"],
        "properties": {
          "displayName": { "type": "string", "maxLength": 128 },
          "nodeType": { "type": "string", "enum": ["GROUP", "REGION", "CITY", "DEALER"] },
          "parentId": { "type": "integer", "format": "int64", "nullable": true },
          "enabled": { "type": "boolean" }
        }
      },
...
      "OrganizationGrant": {
        "type": "object",
        "required": ["organizationNodeId", "includeDescendants"],
        "properties": {
          "organizationNodeId": { "type": "integer", "format": "int64" },
          "includeDescendants": { "type": "boolean" }
        }
      },
      "OrganizationGrantsRequest": {
        "type": "object",
        "required": ["grants"],
        "properties": {
          "grants": {
            "type": "array",
            "uniqueItems": true,
            "items": { "$ref": "#/components/schemas/OrganizationGrant" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/admin/organizations/nodes": {
      "get": {
        "operationId": "listOrganizationNodes",
        "summary": "列出组织树节点（ORGANIZATION_READ）",
        "responses": { "200": { "description": "组织节点列表" }, "403": { "description": "缺少 ORGANIZATION_READ" } }
      },
      "post": {
        "operationId": "createOrganizationNode",
        "summary": "创建组织树节点（ORGANIZATION_MANAGE）",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrganizationNodeRequest" } } }
        },
        "responses": { "200": { "description": "节点已创建" }, "400": { "description": "层级、类型或节点键无效" } }
      }
    },
    "/api/admin/organizations/nodes/{nodeId}": {
      "put": {
        "operationId": "updateOrganizationNode",
        "summary": "更新节点并拒绝组织环或跨根挂载（ORGANIZATION_MANAGE）",
        "parameters": [{ "name": "nodeId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrganizationNodeUpdateRequest" } } }
        },
        "responses": { "200": { "description": "节点已更新" }, "400": { "description": "组织环、跨根挂载、未知或禁用父节点" } }
      }
    },
...
    "/api/admin/organizations/user-grants/{userId}": {
      "put": {
        "operationId": "replaceUserOrganizationGrants",
        "summary": "替换用户组织 grant（ORGANIZATION_GRANT_MANAGE）",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrganizationGrantsRequest" } } } },
        "responses": { "200": { "description": "用户 grant 已替换；下一请求生效" }, "400": { "description": "未知用户、未知或禁用节点" } }
      }
    },
    "/api/admin/organizations/role-grants/{roleId}": {
      "put": {
        "operationId": "replaceRoleOrganizationGrants",
        "summary": "替换角色组织 grant（ORGANIZATION_GRANT_MANAGE）",
        "parameters": [{ "name": "roleId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrganizationGrantsRequest" } } } },
        "responses": { "200": { "description": "角色 grant 已替换；下一请求生效" }, "400": { "description": "未知角色、未知或禁用节点" } }
      }
    },
...
    "/api/reports/drafts": {
      "post": {
        "operationId": "createReportDraft",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ReportDraftRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReportDraftResponse" }
              }
            }
          },
          "400": { "description": "invalid report request" }
        }
      },
      "get": {
        "operationId": "listReportDrafts",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ReportDraftListResponse" } }
            }
          }
        }
      }
    },
    "/api/reports/drafts/{id}": {
      "get": {
        "operationId": "getReportDraft",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReportDraftResponse" } } } },
          "404": { "description": "report draft not found" }
        }
      }
    },
    "/api/reports/drafts/{id}/markdown": {
      "get": {
        "operationId": "exportReportDraftMarkdown",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Markdown report draft", "content": { "text/markdown": { "schema": { "type": "string" } } } },
          "404": { "description": "report draft not found" }
        }
      }
    }
  }
}